Bug#31103386 PERFORMANCE SCHEMA DOES A MEANINGLESS SUMMARY ACCORDING TO DIGEST
Problem

When computing a statement digest, statements
  SELECT * FROM t1 GROUP BY 1;
  SELECT * FROM t1 GROUP BY 2;
are represented with the same digest:
  SELECT * FROM t1 GROUP BY ?;

This puts statistics for potentially very different statements,
that can use different execution plans, into the same digest,
which provides no value.

Likewise for ORDER BY.

Root cause

In SQL, GROUP BY N means to group by the Nth column,
so the literal value present in the syntax (GROUP BY 1)
is not really a literal, it should be treated as a column name.

The rules to reduce literal values into a '?' placeholder
do not account for this SQL syntax specificity.

Fix

Define a new token, TOK_BY_NUMERIC_COLUMN,
to represent a column given by numerical value.

In the digest representation, store the value associated with
a TOK_BY_NUMERIC_COLUMN, similar to TOK_IDENT which has the ident name.

In the digest state, rename m_last_id_index to m_last_peekable_index
for clarity, now that this index applies to both TOK_IDENT and
TOK_BY_NUMERIC_COLUMN.

In sql_yacc.yy, when parsing a GROUP BY or ORDER BY clause,
detect when the expression used is a single numeric token,
and instruct the digest code to handle this special case.

In the digest code, in digest_adjust_by_numeric_column_token(),
adjust the token stream to represent special cases
for ORDER BY and GROUP BY expressions,
by inserting a TOK_BY_NUMERIC_COLUMN with a value in the stream.

In the digest printing code, render TOK_BY_NUMERIC_COLUMN with their value.

As a result:
  SELECT * FROM t1 GROUP BY 1;
  SELECT * FROM t1 GROUP BY 2;
are now represented by two different digests:
  SELECT * FROM t1 GROUP BY 1;
  SELECT * FROM t1 GROUP BY 2;
which is the desired behavior.

Note that any digest previously recorded,
for statements using the GROUP BY N or ORDER BY N syntax,
now has a different digest text and value.

Digests recorded in the firewall component may need to be adjusted.

Change-Id: I635557a72d6042cf0c451f540a8a4e3d46517120