Bug#38402087: Non-equality join predicate prevents hypergraph from using good index
The hypergraph optimizer sometimes did not consider a nested loop join
with a single-row index lookup on the inner side, and ended up using a
hash join that read the entire table. If the table was big, using a
single-row index lookup could be much cheaper.

This happened in queries where a non-equality join predicate
referenced two tables that did not also have an equijoin predicate
between them, and some other join predicate was a sargable equijoin
predicate referencing one of the tables in the non-equality join
predicate.

For example:

    SELECT 1
    FROM t1 JOIN t2 ON t1.x = t2.x
            JOIN t3 ON t2.y = t3.y
    WHERE t1.z <> t3.z

From the equijoin predicates, we get two simple edges in the
hypergraph: t1-t2 and t2-t3. But the inequality in the WHERE clause
cannot be pushed to any of those two edges. In the optimizer trace we see:

    Pushing conditions down.
    - condition (t1.z <> t3.z) makes join reference more relations,
      but could not do anything about it

The simple t1-t2 edge is therefore widened to a t1-{t2,t3} hyperedge,
onto which the inequality can be pushed. (This is an arbitrary choice.
It could also have chosen to widen the t2-t3 edge to {t1,t2}-t3.)

Later, the t1-t2 edge is added back to the hypergraph by
CompleteFullMeshForMultipleEqualities(), so that we have both a simple
edge t1-t2 and a complex hyperedge t1-{t2,t3} connecting t1 and t2.

When trying to join t1 and t2 using the t1-{t2,t3} edge, t2 and t3
must be joined first. Because there is a simple t1-t2 edge,
DisallowParameterizedJoinPath() will refuse to do the t2-t3 join with
an index lookup on t2 parameterized on t1.

When trying to join t1 and t2 using the simple t1-t2 edge, the
optimizer thinks the sargable predicate t1.x = t2.x is redundant with
some other predicate in the join condition, and refuses to use the
index lookup with the following message in the optimizer trace:

    - {EQ_REF, cost=6.48, init_cost=0, rows=1, parm={t1}} has a
      sargable predicate that is redundant with our join predicate,
      skipping

The patch fixes this problem by allowing the inequality to induce a
simple cycle edge between t1 and t3 instead of widening the t1-t2
edge. When there is only a simple t1-t2 edge, the hypergraph optimizer
has no problems using index lookups for the t1.x = t2.x join
predicate.

Change-Id: I40228cbdd7e2fd78b6ac10dc2507da3be047128e