Bug#38351674 ndbmtd-7 pod container mysqlndbcluster restarted with error 2341 THRMAN (Line: 4871)
The ndbinfo.cpustat tables serve data on demand to user queries, supplied
by THRMAN block instances in the datanodes.

THRMAN instances periodically gather OS supplied metrics and data node
timing instrumentation which is summarised and returned when the
cpustat tables are scanned.

The cpustat tables provide a percentage figure for data node cpu
time spent in :
 A Buffer full handling
 B Execution
 C Sending
 D Spinning
 E Sleeping

THRMAN includes clipping code to ensure that the 5 percentages sum
to 100% to be usable, with a self check that causes data node
failure if the 5 percentages exceed 100%.

The clipping code has a flaw in certain cases which can lead
to data node failure :

Case 1 : Overflow on C (Sending)

  if  ((A + B) < 100) AND
      (A + B + C > 100)
  then
    B = 100 - (A+B)
    D = 0
    E = 0

This changes B to be the diff between (A+B) and 100,
likely leaving some extra space for C?
It does not adjust C, so that
  A+B+C+D+E can still be > 100

It is not clear why this transform would be valid.

Case 2 : Overflow on D

  if ((A + B + C) < 100) AND
      (A + B + C + D) > 100)
  then
    B = 100 - (A + B + C)
    E = 0

This changes B to the difference between (A+B+C) and 100,
likely leaving some extra space for D?
It does not adjust D, so that
  A+B+C+D+E can still be > 100

It is not clear why this transform would be valid.

--

It seems like the algorithm intention is to sum up the components,
clipping the first one to exceed 100, and zeroing the remaining.
However it is incorrectly implemented.

It seems simplest to align the algorithm to the apparent
intent, so that the 100% limit is not exceeded.

Change-Id: I480edd88247f20e9422d5f0ebb6bc2c3e69245dc