Bug#38406747 RPM build broken on Oracle Linux 10
RPM builds on modern Linuxes are stricter about RPATH values for
binaries.  Any values containing $ORIGIN should come *before* absolute
paths.

The RPATH values of installed binaries are set according to the cmake
PROPERTY INSTALL_RPATH for each target.

Rewrite the cmake function ADD_INSTALL_RPATH to eliminate duplicate
entries in INSTALL_RPATH, and to always keep is sorted, so that
$ORIGIN entries come before entries starting with '/'.

Router cmake code sets CMAKE_INSTALL_RPATH, and this value will be
used to initialize INSTALL_RPATH for all router targets.  The cmake
list ROUTER_INSTALL_RPATH was identical to CMAKE_INSTALL_RPATH, but is
not needed, so remove it.

Also ensure that all paths in INSTALL_RPATH are stripped of the
trailing '/'.

Change-Id: Ib521e5ff5872c025383393177505f566e88ccd01