WL#17037: Exempt unknown users from connection control's delay criteria
Connection control component/plugin introduces a delay if a number of
failed login attempts for a given authID goes above threshold configured
by the administrator.

This worklog implements a way for the following legal usecase to be
exempted from introducing the delay:
 - connections that do not use MySQL protocol, just probing if the
   server is up and running (for example originating from load balancer)

This commit ports the existing changes from connection_control
component to the connection_control plugin.

There are following differences compared to component code:
- system variable renamed from
  "component_connection_control.exempt_unknown_users" to
  "connection_control_exempt_unknown_users"
- status variable renamed from
 "Component_connection_control_exempted_unknown_users" to
 "Connection_control_exempted_unknown_users"

Change-Id: I189125f423c3a91df1096f344cd943df20f53964