Bug#38418831: Hypergraph: Wrong result when sorting hash join result reading from temptable
Wrong results were seen when executing some query plans that had a
SORT access path on top of a HASH_JOIN access path, with no
materialization step in between. This happened if the SORT needed row
IDs from the underlying tables and the HASH_JOIN read from a
materialized derived table that was so big it had overflown to disk.

The results were wrong because the sorting iterator set up the sort
buffer before the derived table had been materialized, and therefore
did not know that it was too big to fit in an in-memory table. Since
the row ID size (handler::ref_length) was different in a memory-based
temporary table than in a disk-based temporary table, the rows written
to the sort buffer did not match the row layout in the buffer, and the
contents of the sort buffer were corrupted.

This was caused by the fix for bug#37746132, which made hash join
iterator initialize its child iterators lazily on the first call to
Read(). Even though the sorting iterator initialized the hash join
iterator below it before it set up the sort buffer, the iterators
below the hash join iterator were not initialized, and the derived
table was not materialized until later.

The lazy initialization was introduced in order to prevent that the
first row read from the probe table during initialization was modified
by another iterator higher up between initialization and the first
call to Read() on the iterator.

This patch reverts the original fix so that hash join iterator
performs eager initialization, as before. To avoid reintroducing
bug#37746132, the row read from the probe table during initialization
is copied to a dedicated buffer, from which it is restored on the
first call to Read().

Additionally, while at it, the calls to PrepareForRequestRowId() on
m_build_input_tables were moved so that only a single caller was
needed.

Change-Id: Ia48d14ac86d69ff2ac88115fb36b1e6a3fb02b92