Bug#37953883 Table_map entry is logged when sql_log_bin=0 [addendum]
Fixes a bad binlog signalling, i.e., an event with anyValue !=0 but
with no clear special word, that would cause data events to not be
logged.

Issue was a non-existent initialization of the m_any_value on the
NdbOperation that ultimately produced that data event, since it wasn't
used as it is now used ( OF_FLAGS protected the assignment, previously,
but with NdbRecAttr API these cannot be used, so the NdbBlob handle
must rely on m_any_value != 0 ).

Change-Id: I16c620f8b2e7dbf6535adfd1e3c87df1048cc810