Bug#38500023 SUMA sends BLOB data with NO-LOGGING set
Problem:

When a table with BLOB columns is setup for binary logging, it is
expected that any do-not-log data (sql_log_bin=0) is not received by
the binlog injector since the no-logging filter is by default set to
all tables. However, the binary injector is receiving such data events
as verified by the NdbApi counters
`Ndb_api_event_bytes_count_injector`.

Solution:

When the NdbEventOperation for the table is setup, the additionally
created NdbEventOperation's for the BLOB columns must contain the same
parameterization of the Main operation.

Change-Id: I26ba21c75019d29e23c3b0f81e2e03dd9d823092