Bug#38262728: mysqld fails when installing component_test_event_tracker_producer_a
The event handling code was assuming that there are certain extra data available in the THD (event_information) or that certain members of the even data are filled in properly.
This might not always be the case.
Added runtime hardening checks that generate a warning and do nothing more.
Added a test case with a test component loaded out of context.

Change-Id: I5aa6ae4d57bf63c6838cae3e04c95b9ab019d274