Bug#38450918 - Application crashes during async task execution
Description
===========
This incorrect IAT entry is caused by a misconfiguration in `router/src/harness/src/CMakeLists.txt`,
specifically in the `MERGE_LIBRARIES_SHARED(router_mysqlclient...)` section. The `client_errors` symbol
is listed as:
   ?client_errors@@3PAPEBDA

However, it should be specified as:
   ?client_errors@@3PAPEBDA DATA

This causes that importer dll tries to dereference the IAT entry as data pointer, which is wrong.

Fix
===
Switch the usage from the internal symbol client_errors to my_get_err_msg.
The case is covered by the unstable MTR test router.mysql_tasks,
but it is only started for cloud builds. To fix the test or write a new one,
the following bug was created:

    Bug#38456691 - i_router.mysql_task runs only on linux/cloud package

Change-Id: I1dff8da3211515bc8934413ca8184b1a3b3b13a4