Bug#38235957: failure to DROP TABLE with an invalid CONSTRAINT
SQL in table definitions may or may not be valid depending
on the sql_mode. When invalid, we should process this
correctly as the inability to open the table in question,
rather than to imply that the table does not exist.
If the file exists, mysqld, InnoDB, and the file system
should all agree on this.

Change-Id: I7e22ef49db1847b3ff86bee9556b8ab6281158cb