Bug#38380392: 8.0.43 GR hangup PRIMARY with hung 'waiting for handler commit'               when 2 of 6 nodes unreachable
Description:
------------
A Group Replication InnoDB Cluster configured with paxos_single_leader=ON
became unwritable. When 2 nodes of a 6-node geographically dispersed InnoDB
Cluster lost connectivity, the cluster PRIMARY stopped accepting writes,
and COMMIT/INSERT operations hung.

Analysis:
------------
Scenario:
Consider Group Replication(GR) clustor with three nodes
group_replication_paxos_single_leader=ON;
   Node0(Primary), Node1(Secondary1), Node2(Secondary2).

The issue involving a long-running transaction stuck in the
`waiting for handler commit` state unfolds as follows:
On Primary:
 1. A transaction(TXN) is initiated.
 2. The TXN message is pushed to the XCom proposer queue. The connection
    thread waits on `group_replication_trans_before_commit`.
 3. The proposer picks up the TXN message, acquires a site (SITE0 – where
    Node0 is the leader), assigns a synod number (SYN01), and proposes it.
 4. The proposer waits for the message to be accepted.
 5. The proposer gets the message accepted, but later detects a mismatch
    between the proposed message and the learned message.
    The reason for this mismatch is due to the intermittent disconnection
    between the primary and the secondaries. This scenario is expected.
 6. The proposer re-proposes the TXN and waits for it to be accepted:
    6.a A new synod number (SYN02) is reserved for the TXN,
        with SITE0 (Node0 as leader).
    6.b The proposer yields the CPU during the `while (too_far(*msgno))`
        loop in the reserve_synod_number function.
    6.c The executor gets scheduled, processes a `view_msg` message,
        and generates a new site (SITE1 – Node1 becomes the leader).
    6.d The proposer resumes, assigns SYN02 to the TXN message,
        and proposes it under SITE0.
    6.e The proposer waits for the TXN message to be accepted, but no
        `ack_accept_op` is received from the other nodes since the proposal
        is still tied to the past SITE0 configuration.
    6.f The proposer continues waiting indefinitely in the loop.
 7. As a result, no message is sent back to GCS/GR, and the transaction
    remains indefinitely stuck on `group_replication_trans_before_commit`.

Fix:
----
Proposer task updates site after resume in `reserve_synod_number`.

Previously, the site was updated before yielding the CPU in the
`reserve_synod_number` function. If a site update occurred while the proposer
was yielding, the proposer could continue using an outdated site configuration
when proposing a message.

With this change, the site is updated after the proposer resumes, ensuring that
it always attaches the latest site information to the message before proposing.
If a site update occurs while the proposer is yielding, the outdated message
will be ignored and a re-reservation of the synod number will be triggered.
This prevents the proposer from getting stuck in a loop due to proposals tied
to past site configurations.

Change-Id: Ia82ecaf05b5893b08d255d67230d3875e6603238