Bug#38422448 Data Node crash DBACC (Line: 7516) 0x00000000 Check oprec_pool.getValidPtr(tatrOpPtr) failed
Problem

A crash is observed in the ACC takeover code execACC_TO_REQ(),
where the following validation fails :

  ndbrequire(oprec_pool.getValidPtr(tatrOpPtr));

This is checking that the scan lock ACC operation reference
is valid in the pool.

This reference has previously been checked for validity in
execACCKEYREQ(), where an invalid reference should result
in a takeover error on the operation resulting in transaction
rollback, rather than data node failure.

However the execACCKEYREQ() check is incorrect.

This check is fixed.

Change-Id: Icdf94403ba81b69d1be75b34e60bdbd60b5f91f7