Bug#38458235 binary library export
Contribution http://bugs.mysql.com/file.php?id=34322

When dumping the binary library code, we can base64 encode it, or hex encode it.
When dumping in hex mode, we want to dump each character as a hex code.
In order for std::ostream to display it as a number but not a character, we cast every character to an integer.
Unfortunately static_cast changes the binary representation of the objects that are cast.
Since "char" is a signed data type (on many platforms), the characters with ASCII code above 127 are treated as negative numbers. When cast to a bigger number, the signage is preserved. E.g. static_cast<int>(char{-1}) is cast to "-1" in integer form. But binary they differ - the -1 in 8-bit form is "0xFF". But -1 in 32 bit integer form is "0xFFFFFF".
As a result, characters with ASCII > 127 are output with extra, errorneous preceding FF's.

In order to properly cast "char" to "unsigned int", it has to be done in two steps:
 * static_cast<unsigned char> - first remove the signage but preserves the underlying data. E.g. cast -1 to 255.
 * static_cast<unsigned int> - so it can be displayed as integer (in hex form) rather than character. But keep 255 as 255 - and display it as "FF".

Note: The same effect can be achieve using the unsafer reinterpret_cast - this cast does not change the underlying data.

Change-Id: I0e8cb89e50724615ec98dd93576ad3d502a942d4