BUG#38472467: Log error message before raising the diagnostics area !is_set() assertion
* Background

The server allows itself to report at most one error to a client
session, and requires that after reporting an error, it returns an
error status to the client. If any of that is violated, it is a bug,
and the server raises an assertion in debug mode.

It is a common mistake to introduce bugs that cause this assert to be
raised (search for bugs with "!is_set" in their titles). The reason
for the bug is usually that an error is reported using my_error, but
the function forgot to return an error status to the caller (or some
of the function higher up on the stack forgot to propagate the error
status).

* Problem

When such bugs occur, the analysis usually involves finding the site
that reported the previous error, because that's where the return
status propagation was missing. The site where the error was reported
is not tracked, but even just the message is often enough.
Unfortunately, the message can only be recovered from a debugger
session, which is not always possible (especially if the bug is hard
to reproduce). That makes it unnecessarily hard to analyze and fix
such bugs.

* Fix

Before raising the assertion, write an error to the log, containing
the previous status, the requested status, the previous error message,
and the requested error message. This makes it easier for developers
to find what the previous site that reported the error might have
been.

Change-Id: Ia0334e3f14b507504da33cc3aa8c4545f7553a01