Bug#38151265 Assertion Failure in Backup::execDEFINE_BACKUP_REQ 5/N
Problem:
A BackupRecord in a participant node is not released from the
pool when a backup is aborted. A subsequent backup operation
fails when attempting to seize it.

The sequence of events that leads to the crash is as follows:
1. Master receives a BACKUP_FRAGMENT_REF signal from a participant.
2. Master sends ABORT_BACKUP_ORD with flag ABORT_SCAN to all worker
threads that it thinks were scanning a fragment at that moment.
3. When the ABORT_BACKUP_ORD arrives to the participant one of two
things happens:
a) If the scan is still running then it will stop soon and a
BACKUP_FRAGMENT_REF/CONF is sent to the Master.
b) If the participant was already finished scanning when the
ABORT_BACKUP_ORD(AbortScan) arrives then it will be idle, and the
effect will be: Set error + Change ptr.p->m_gsn to ABORT_BACKUP_ORD.
This is a problem for a subsequent Master failure as a participant in
that state maps to 'do nothing' and so backup resources are not
cleaned.

Solution:
change the execABORT_BACKUP_ORD(AbortScan) case to only change the
ptr.p->m_gsn to ABORT_BACKUP_ORD if there is a scan active in the
participant. (e.g. if ptr.p->m_gsn was BACKUP_FRAGMENT_REQ).
If it was BACKUP_FRAGMENT_REF or BACKUP_FRAGMENT_CONF, ptr.p->m_gsn
is kept, so a subsequent Master failure will effectively clean
the backup resources.

Change-Id: I24dee069f1c7f635b4b7590378da0555bdb86fae