Bug#38441744 Hypergraph: Too much time spent in GetReadSetWidth() when planning query
Post-push fix:

The fix for the bug introduced the new member variable
m_read_set_width in the TABLE struct. It is initialized lazily when
TABLE::read_set_width() is called, but it is never reset. Since TABLE
objects are recycled and could be used again for a different table, or
for the same table with a different read set, the member variable must
be reset before the TABLE object is used again, otherwise
TABLE::read_set_width() may return a stale value. This caused test
failures.

Fixed by caching the read set width in the optimizer-internal
JoinHypergraph::Node class instead of in the server-wide TABLE class.
Since the JoinHypergraph::Node objects are not recycled and reused, we
avoid the risk of keeping stale data between statements.

An additional advantage is that GetJoinRowWidth() can iterate over the
bits in the NodeMap of the subplan instead of walking the AccessPath
tree to find the tables in the subplan. Iterating over the bits in
NodeMap is faster, so this shaves off another 6% of the planning time
for the query in the bug report.

Change-Id: Id9a898a957daef4c71e0e1b41c936f1c00426b8a