Bug#38439704: SIG6 with hypergraph optimizer - assertion failure in RedundantThroughSargable at join_optimizer.cc
An assertion failure was seen in some queries with a large number of
join predicates (so large that the OverflowBitsets pointing into
JoinHypergraph::predicates could not be inlined), one of the joins had
at least two non-equijoin predicates, and there was at least one
sargable join predicate.

It was caused by a loop over the non-equijoin predicates in
CacheCostInfoForJoinConditions(), which forgot to increment one of the
iterators and therefore kept updating the first element of an array
instead of updating each element in the array.

Fixed by adding the missing increment operation.

Change-Id: Ic9538ad61b3542a9aa75125c89196477127944a7