Bug#38440648: Assertion `false && "Inconsistent row counts for different AccessPath objects."' failed.
Some queries got assertion failures due to inconsistent row estimates
when using the hypergraph optimizer.

This was a regression after bug#36714667. The fix for that bug moved
the selectivity estimation of join predicates later. It also intended
to move the sorting of the join predicates on selectivity later, since
it cannot be done until the selectivity has been calculated. It did
add a new call to SortPredicates() after the selectivity calculation,
but it forgot to remove the earlier call which now happens before the
calculation, so the sorting was done twice. In addition to being
unnecessary extra work, the first sorting meant that the predicates
were seen in a different order than expected in
EstimateJoinConditionSelectivities(), and the selectivities were
assigned to the wrong predicates. This caused inconsistencies in the
row estimates later.

This patch adds more assertions to
EstimateJoinConditionSelectivities() to detect earlier if the
selectivity is assigned to the wrong predicate. And it removes the
unnecessary call to SortPredicates() in PromoteCycleJoinPredicates(),
so that the join predicates are sorted only in
EstimateJoinConditionSelectivities().

Change-Id: I921aaaecfb59205f75154c458a05aaf9b8c5fdaa