Bug#36714667: EstimateEqualPredicateSelectivity is called several times for same join cond [post-push fix]
The original fix for this bug moved the join condition selectivity
estimation later, but the code that wrote the header for the join
condition stayed at the old location. This led to confusing tracing of
join condition selectivities.

For example, the trace for the query

    SELECT * FROM t1, t2, t3 WHERE t1.x = t2.x AND t2.y = t3.y

could contain a section that looked like this:

    Selectivity of join (t1.x = t2.x):
    Selectivity of join (t2.y = t3.y):
     - fallback selectivity for multiple equal(t1.x, t2.x) = 0.1
     - fallback selectivity for multiple equal(t2.y, t3.y) = 0.1

This patch moves the printing of the header to just before the join
condition selectivity is calculated, so that the above trace changes
to this:

    Selectivity of join (t1.x = t2.x):
     - fallback selectivity for multiple equal(t1.x, t2.x) = 0.1
    Selectivity of join (t2.y = t3.y):
     - fallback selectivity for multiple equal(t2.y, t3.y) = 0.1

Change-Id: Ifc7b56d660b78294a658e74291eb82436abb4eaa