BUG#28590993: MySQL generates GTID gaps because of option replica-skip-errors [8.4 + trunk]
* Background

When the replica applier completes processing a transaction and the
applier does not stop as a result of an error, it must "advance the
position". When using GTIDs, the position is represented in
gtid_executed. Therefore, it has to log the GTID even for transactions
that were filtered-out or for any other reason did not result in any
write to the replica binary log. To this end, it generates empty
transactions.

* Problem

When using replica-skip-errors, and a DDL produced a suppressed error,
the replica would not generate an empty transaction, thus did not
"advance the position". The result is that the replica's gtid_executed
will have "gaps", i.e., multiple non-contiguous intervals, and that it
will try to fetch the same transaction in future reconnect attempts,
which eventually will fail when the source purges the transaction.

* Analysis

Normally, the server writes empty transactions at the end of
successful statements which hold ownership of a GTID but did not
produce any contents in the binary log. But two conditions prevented
this logic from being invoked on replicas using replica-skip-errors:

1. A DDL that fails will rollback the transaction, which releases
   ownership of the GTID.

2. The point where empty GTID transactions are usually generated
   occurs before the point where replica-skip-errors overrides the
   error state with a success state. Therefore, the transaction is
   still in an error state when it checks whether an empty GTID
   transactions is needed. The error prevents it from generating an
   empty transaction.

* Solution

1. In the function that releases GTID ownership, suppress the GTID
   ownership release in case the session has an error that should be
   ignored.

   To keep the code that releases GTID ownership independent on error
   suppressions, we implement this as follows:

   - Introduce an API to register callable objects that determine if
     GTID ownership release should be suppressed.

   - Augment the GTID ownership release code to check any registered
     functions.

   - When applying a Query_log_event while there are suppressed
     errors, register such a function which checks if the session has
     an error state with an error code that is to be ignored.

2. Invoke the code that generates an empty transaction if needed, at
   the end of processing a Query_log_event, if the thread is not being
   stopped and we are not in a transaction context.

Change-Id: I76ff488eefe63eb3be7a385610c8c2e6a189123e