Bug#38488470 AddressSanitizer: heap-use-after-free in pfs_end_statement_vc
This is a follow up after:
  BUG 36445818 P_S doesn't record SQL or digest
  for prepared statements for Non-mysql client

Problem

The asan build reports heap-use-after-free errors,
when using the digest query text for a DEALLOCATE PREPARE statement.

Root cause

Prepared_statement::psi_instrumentation()
returns data stored in the prepared statement,
which is about to be destroyed.

Fix

In Prepared_statement::psi_instrumentation(),
make a copy of the query text,
when instrumenting a DEALLOCATE PREPARE.

Change-Id: I4b3d6b0188d845661caaa73d4b58915b1eef39ba