BUG#38494914: Test rpl_gtid.rpl_replica_invalid_time_zone fails when executed twice
* Background

The test rpl_gtid.rpl_replica_invalid_time_zone simulates a situation
where a timezone from a replicated event does not exist in the
replica's timezone database. It simulates the situation by renaming
the `mysql.time_zone` table so that the server is effectively unaware
of any timezones. Then it verifies that the proper error is generated
when trying to apply the event containing the timezone informration.

* Problem

On the first use of the timezone database, it is cached in
memory. Therefore, on the second run of the test on the same server
instance, it does not query the table but uses the cached timezones,
which at that time exist. Therefore, contrary to the test expectation,
it finds the timezone and the test fails.

* Fix

Restart the server in the beginning of the test

Change-Id: I96b3891fd28ffe5433a2440eec1653049fe228e9