Bug#38487373: Assertion failed: time.time_type == MYSQL_TIMESTAMP_TIME
The problem here is an assert in eval_temporal(), when evaluating
a string argument. The argument is checked for being a TIME or a
DATE/DATETIME value, by evaluating the string in functions
str_to_time() and str_to_datetime(). The string in question is
-1974-10-18 08:57:35.681107, which is an invalid datetime since we do
not accept negative year values. However, str_to_time() classifies
this as a valid datetime value, whereas str_to_datetime() does not,
thus we have an obvious inconsistency.

The fix is to modify str_to_time() so that it reports negative
datetime values as errors.

Change-Id: I7967ab6070c774e182b4d27a7844df51690fa345