Bug#38499119 AddressSanitizer: heap-use-after-free at pfs_end_statement_vc (2)
This is a second follow up after:
  BUG 36445818 P_S doesn't record SQL or digest
  for prepared statements for Non-mysql client

Problem

(1)

The asan build reports heap-use-after-free errors,
when using the query text for a PREPARE statement.

(2)

The asan build reports memcpy-param-overlap errors,
when executing sql_digest_storage::prefix_and_copy
for a PREPARE statement.

Root cause

(1)

The code uses the display query text
from the prepared statement.

When a PREPARE statement later fails,
the prepared statement is destroyed,
leading to the issue.

(2)

During PREPARE:

The statement digest from PREPARE is copied
into the prepared statement.

Later the digest is copied back into the PREPARE
statement with a prefix.

This prefix_and_copy copies the digest in the
PREPARE statement into itself,
instead of using the prepared statement as source.

Fix

(1)

Make a copy of the display query text,
to account for failure paths.

(2)

Call prefix_and_copy using the prepared statement digest,
so that source and destination do not overlap.

Change-Id: Iffcd7ba7b9603003439b0119aaf544cbccd7f0a4