Bug#37876784:  Creating PFS table while shutdown causes component is unusable
When registering the PFS table into the DD, onlky a table lock was
taken.
This needs to be extended to cover other MDL locks too, so that nobody
can alter the dictionary in a way that will prevent the table from being
added.

Fixed by "borrowing" the NDB sequence that registers tables.

No test case, since it's time-sensitive. Tested manually.

Change-Id: I5f7182d77d135ba912a4fdbc66c88c1e1c07b43d