Bug#38515974  AT: testNodeRestart -n ChangeNumLDMsNR is wrongly skipped in 4node4rpl
Problem:
Test is intended to run only when data node is multi threaded (ndbmtd)
To accomplish that, test queries the ndbinfo table 'threads' to count
number of execution threads, if that number is 1 test is skipped.

Probmem is that, in 4node4rpl configuration the following configs are
set:

AutomaticThreadConfig=1
NumCPUs=1

As consequence data node will run with only one execution thread.
This causes test to think it is a single threaded ndbd data node
and the test is skipped.

Solution:
When querying 'threads' ndbinfo table, thread_name column is checked,
if there is only one thread and its name is 'main' data node is
multi threaded, if name is 'main_rep_recv' data node is single
threaded. This information is used to skip test only when data node
is ndbd.

Change-Id: I8fc864b3c13d55441b622878c55acaebf0479247