Bug#11746738: UNCLEAR ERROR MESSAGE WHEN CREATE USER FAILS DUE TO DUPLICATE KEY
Added an option to print a reason (in parenthesis, in English) for every user account (role or login) failing in ACL statements (CREATE, DROP etc).

Recorded *some* of the result files, to give a taste of what these looks like. Will follow up with a comple re-record once we settle on the functionality.

Change-Id: Ibec559e9b2c84252eb13e1fc46770b17e88c0c0f