Bug#22240504: MYSQLDUMP --ROUTINES MISQUOTES DATABASE NAMES
The problem was that the string quoted to go into a quoted string (single quote-delimited) was used into identifier quotes (backtick-limited).
In the backtick quotes backslash is not a special case so no backslash escaping was necessary, yet it was being done.

Fixed the --routines and --events to properly quote the dabase name.
Test case added.

Change-Id: I0b6868eba69665b11f65eff0b69f8b92c92e44ec