Bug#38531908 Test failure in main.subquery_to_derived_hypergraph
Problem

Test main.subquery_to_derived_hypergraph fails
Test main.subquery_scalar_to_derived fails

Root cause

The number of digests in the test suite
was increased from 200 to 400,
with the following change:

  Bug 38523909 Misc test failures involving digests in --ps-protocol

These tests outputs the optimizer test plan,
which depends on the number of digests.

Fix

Restore loose-performance-schema-digests-size=200
for these tests, so that the test conditions are unchanged.

Change-Id: I1df3fe5794e1e9e33b1fa5470059aa3cd53d3e20