Bug#28987279: infinite loop when csv engine has no permission to meta file
A contribution by Yan Huang

When the meta file for ha_tina could not be opened (because of e.g.
 permission issues), opening the table goes into an infinite loop.

This is happening because:
- open fails
- the table is marked as crashed
- an automatic repair is attempted
- the repair only checks the data file and succeeds
- table is considered repaird
- open is tried again (go to start).

On top of that close was trying to write to the meta file and was
causing an dbug assert when the meta file is not opened.

Fixed by:

1. Making sure close() can succeed, even with a missing meta file.
2. Making sure repair() doesn't succeed if it can't open the meta file.

Change-Id: I7e634d667d6cd609b943955da39a9e49f26eb3cf