Bug#38032670 MTR falsely marking tests as skipped with --hypergraph
Usage of the --hypergraph option in MTR could result in diffs
which are expected and ignored. Tests with such cases were
marked as 'skipped' with a corresponding reason. This could
cause ambiguity between actually skipped tests and those tests
which pass when result sections with expected diffs are ignored.

The patch updates this behavior so that MTR marks such tests
optionally/provisionally passed ('[ opt-pass ]' on console) in
order to not mix such outcomes with skipped tests.

Change-Id: Idec910465aaf150ee20c514d3243d128aca97f75