Bug#26750972: TRAILING # COMMENT ON !INCLUDEDIR / !INCLUDE DIRECTIVE CAUSES ERROR
The argument parser for !include and !includedir just took everything
after the command itself and the whitespace after it for
a file/directory name.
It did not recognize or process quotes and comments as it's done for the
other lines.

Fixed the parser for the argument of !include and !includedir to support
trailing comments (!include foo # comment). A trailing comment is
considered the symbols after the first # sign starting from the
argument's start.

Test case added.

Change-Id: Ib9f4d777e788cdd2d7f23735d65a06bb1d0fd82e